Move useQuery under hooks/query folder
