+++ packages/frontend/src/pages/Landing/index.tsx
@@ -5,7 +5,6 @@ import { useTranslation } from 'react-i18next';
 import { ContainerWrapper, SubtitleWrapper, ParagraphWrapper } from './index.styles';
 import { ButtonWrapper } from '../../styles/ButtonWrapper';
 import { Footer } from '../../components/Footer';
-import { useQuery } from '@tanstack/react-query';
 
 export const Landing = () => {
   const { t } = useTranslation(['common', 'landing']);
@@ -37,16 +36,6 @@ export const Landing = () => {
   };
 
   const LandingView = () => {
-    useQuery({
-      queryKey: ['dummy2'],
-      queryFn: async () => {
-        const response = await fetch('/api/test');
-        if (!response.ok) {
-          throw new Error('Dummy2 failed');
-        }
-        return response.json();
-      },
-    });
     return (
       <>
         <SubtitleWrapper variant="subtitle1">{t('landing:welcome.text')}</SubtitleWrapper>
